requirements.txt
Flask
selenium
webdriver-manager
beautifulsoup4
pandas
python-dotenv
